# openapi (2).json
{
  "openapi": "3.0.0",
  "info": {
    "x-api-id": "7f7ceff3-627d-4c9d-8333-069dbd112bb0",
    "x-audience": "external-public",
    "version": "1.0.0",
    "title": "Trademark search",
    "x-ibm-name": "trademark-search",
    "x-pathalias": "trademark-search",
    "description": "Perform searches within EUIPO's database and retrieve all available information on a given trade mark.\n",
    "contact": {
      "name": "API Platform Team",
      "url": "https://dev.euipo.europa.eu",
      "email": "apiplatform@euipo.europa.eu"
    },
    "termsOfService": "https://euipo.europa.eu/tunnel-web/secure/webdav/guest/document_library/contentPdfs/law_and_practice/decisions_president/EX-22-05_en_annex_I_version_2_en.pdf",
    "x-ip-domain": [
      "Trademark"
    ],
    "x-service-type": [
      "Search",
      "Dissemination"
    ],
    "x-content-type": [
      "Bibliographic data",
      "Legal status",
      "Classification"
    ],
    "x-product-page-url": "https://dev.euipo.europa.eu/product/trademark-search_100",
    "x-ibm-summary": ""
  },
  "servers": [
    {
      "url": "https://api-sandbox.euipo.europa.eu/trademark-search"
    }
  ],
...
  "security": [
    {
      "ClientID": [],
      "Oauth2ClientCredentials": [
        "uid"
      ]
    },
    {
      "ClientID": [],
      "Oauth2AuthorizationCode": [
        "trademark-search.trademarks.read"
      ]
    }
  ],
  "paths": {
    "/trademarks": {
      "get": {
        "summary": "Search for trade marks",
        "description": "Search for trade marks that match the criteria below. Results are returned in pages (check parameters page and size for more details).",
        "operationId": "searchTrademarks",
        "tags": [
          "Trademarks"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Fields"
          },
          {
            "$ref": "#/components/parameters/Query"
          },
          {
            "$ref": "#/components/parameters/PageNumber"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "$ref": "#/components/parameters/Sort"
          }
        ],
        "responses": {
          "200": {
            "description": "Result of the search",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrademarkSearchResult"
                }
              }
            },
            "headers": {
              "X-RateLimit-Limit": {
                "$ref": "#/components/headers/X-RateLimit-Limit"
              },
              "X-RateLimit-Remaining": {
                "$ref": "#/components/headers/X-RateLimit-Remaining"
              },
              "X-RateLimit-Reset": {
                "$ref": "#/components/headers/X-RateLimit-Reset"
              },
              "Retry-After": {
                "$ref": "#/components/headers/Retry-After"
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/SearchTrademarkBadRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitError"
          },
          "default": {
            "$ref": "#/components/responses/DefaultError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ClientID": {
        "type": "apiKey",
        "in": "header",
        "name": "X-IBM-Client-Id"
      },
      "Oauth2ClientCredentials": {
        "type": "oauth2",
        "description": "This API operates using two different OAuth2 flows, each offering a unique way of handling user data and trade mark \ninformation:\n- **Client Credentials Flow:** In this method, the final user remains anonymous. The API authenticates the client \nbased on their credentials but does not have access to the user’s identity. As a result, under certain conditions, \nthe API may choose to hide specific parts of the trade mark information to maintain privacy and security.\n",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://auth-sandbox.euipo.europa.eu/oidc/accessToken",
            "scopes": {
              "uid": "Grants read access to trade marks"
            }
          }
        },
        "x-ibm-oauth-provider": "sandbox-oauth"
      },
      "Oauth2AuthorizationCode": {
        "type": "oauth2",
        "description": "- **Authorization Code Flow:** This flow involves the identification of the final user. Since the user is authenticated \nand their identity is known, the API can provide complete information about the trade mark. This method allows \nfor a more personalized user experience, as it tailors the data access according to the user’s identity.\n",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth-sandbox.euipo.europa.eu/oidc/authorize",
            "tokenUrl": "https://auth-sandbox.euipo.europa.eu/oidc/accessToken",
            "refreshUrl": "https://auth-sandbox.euipo.europa.eu/oidc/accessToken",
            "scopes": {
              "trademark-search.trademarks.read": "Grants read access to trade marks"
            }
          }
        },
        "x-ibm-oauth-provider": "sandbox-oauth"
      }
    },
    "headers": {
      "X-RateLimit-Limit": {
        "schema": {
          "type": "integer"
        },
        "description": "The maximum number of requests you're permitted to make per time period."
      },
      "X-RateLimit-Remaining": {
        "schema": {
          "type": "integer"
        },
        "description": "The number of requests remaining in the current rate limit window."
      },
      "X-RateLimit-Reset": {
        "schema": {
          "type": "integer"
        },
        "description": "The relative time in seconds until the current rate-limit window resets.\n"
      },
      "Retry-After": {
        "schema": {
          "type": "integer"
        },
        "description": "The number of seconds to wait until the rate limit window resets. Only sent when the rate limit is reached."
      }
    },
    "parameters": {
      "PageNumber": {
        "in": "query",
        "name": "page",
        "schema": {
          "$ref": "#/components/schemas/PageNumber"
        },
        "description": "The page to be returned"
      },
      "PageSize": {
        "in": "query",
        "name": "size",
        "schema": {
          "$ref": "#/components/schemas/PageSize"
        },
        "description": "The number of elements to be returned"
      },
      "Sort": {
        "in": "query",
        "name": "sort",
        "schema": {
          "$ref": "#/components/schemas/Sort"
        },
        "description": "The sort query parameter determines the sort order. It is the sort field that uses the ':' separator\nplus the value 'asc' (ascending) or 'desc' (descending) to indicate sorting direction.\n\n ### List of sortable fields\n | Field                               |\n |-------------------------------------|\n | applicationNumber                   |\n | applicationReference                |\n | applicationDate                     |\n | designationDate                     |\n | registrationDate                    |\n | markFeature                         |\n | wordMarkSpecification.verbalElement |\n | updateDate                          |\n",
        "example": "applicationNumber:asc"
      },
...
      "Fields": {
        "in": "query",
        "name": "fields",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 1000
        },
        "description": "The `fields` query parameter determines the fields returned with the response payload object.\n\nThe syntax of the query `fields` value is defined by the following [EBNF](https://en.wikipedia.org/wiki/Extended_Backus%E2%80%93Naur_form) grammar.\n```\nfilter = filter-item, {\",\", filter};\nfilter-item = all | field-name | field-name, negation, \"(\" field-names \")\" | field-name, \"(\", filter-item, {\",\", filter-item}, \")\";\nfield-names = field-name, {\",\", field-name};\nfield-name = {letter-digit}-;\nletter-digit = ? alphanumeric character ?;\nnegation = \"!\";\nall = \"*\";\n```\n\n**Examples of valid `fields` values:**\n- Return trademarks root object with only the applicationNumber field:<br>\n```trademarks(applicationNumber)```\n\n- Return trademarks root object with the applicationNumber field and the nested applicants object with only its\n  identifier field:<br>\n```trademarks(applicationNumber,applicants(identifier))```\n\n- Return trademarks root object excluding applicantReference field:<br>\n```trademarks!(applicantReference)```\n\n- Return `trademarks` root object with the `applicationNumber` field and the nested `applicants` object with all its\nfields but `name`:<br>\n```trademarks(applicationNumber,applicants!(name))```\n\n- Return root object without `goodsAndServices` field:<br>\n```!(goodsAndServices)```\n\n- Return root object only with `applicationNumber` field:<br>\n```applicationNumber```\n\n- Return root object excluding `description` field from 'goodsAndServices' field:<br>\n```goodsAndServices!(description),*```\n",
        "example": "!(goodsAndServices)"
      },
      "Query": {
        "in": "query",
        "name": "query",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 5000
        },
        "description": "The `query` parameter allows to search for a specific set of trademarks using RSQL expressions.\nRSQL is a query language for parametrized filtering of entries in RESTful APIs. It uses an URI-friendly syntax,\nthere are no unsafe characters, so URL encoding is not required.\n\nThe syntax of RQSL is defined by the following [EBNF](https://en.wikipedia.org/wiki/Extended_Backus%E2%80%93Naur_form) grammar.\n```\ninput          = or, EOF;\nor             = and, { \",\" , and };\nand            = constraint, { \";\" , constraint };\nconstraint     = ( group | comparison );\ngroup          = \"(\", or, \")\";\n\ncomparison     = selector, comparison-op, arguments;\n\nselector       = unreserved-str;\n\ncomparison-op  = comp-fiql | comp-alt;\ncomp-fiql      = ( ( \"=\", { ALPHA } ) | \"!\" ), \"=\";\ncomp-alt       = ( \">\" | \"<\" ), [ \"=\" ];\n\narguments      = ( \"(\", value, { \",\" , value }, \")\" ) | value;\nvalue          = unreserved-str | double-quoted | single-quoted;\n\nunreserved-str = unreserved, { unreserved }\nsingle-quoted  = \"'\", { ( escaped | all-chars - ( \"'\" | \"\\\" ) ) }, \"'\";\ndouble-quoted  = '\"', { ( escaped | all-chars - ( '\"' | \"\\\" ) ) }, '\"';\n\nreserved       = '\"' | \"'\" | \"(\" | \")\" | \";\" | \",\" | \"=\" | \"!\" | \"~\" | \"<\" | \">\";\nunreserved     = all-chars - reserved - \" \";\nescaped        = \"\\\", all-chars;\nall-chars      = ? all unicode characters ?;\n```\n\n### Available comparison operators\n\n| Operator | Alternative Syntax | Meaning                  |\n|----------|--------------------|--------------------------|\n| ==       |                    | Equal to                 |\n| !=       |                    | Not equal to             |\n| <        | =lt=               | Less than                |\n| <=       | =le=               | Less than or equal       |\n| >        | =gt=               | Greater than operator    |\n| >=       | =ge=               | Greater than or equal to |\n| =in=     |                    | In                       |\n| =out=    |                    | Not in                   |\n| =all=    |                    | All in                   |\n\n*An asterisk can be used as a wildcard with both the equal and not equal operators*\n\n### Available logical operators\n| Operator | Meaning     |\n|----------|-------------|\n| or       | Logical OR  |\n| and      | Logical AND |\n\n### Supported Fields\n| Field                               | Supported operators                        |\n|-------------------------------------|--------------------------------------------|\n| applicationNumber                   | \"==\", \"!=\"                                 |\n| applicantReference                  | \"==\", \"!=\"                                 |\n| applicationLanguage                 | \"==\", \"!=\"                                 |\n| secondLanguage                      | \"==\", \"!=\"                                 |\n| markKind                            | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| markFeature                         | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| markBasis                           | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| wordMarkSpecification.verbalElement | \"==\", \"!=\"                                 |\n| applicants.identifier               | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| applicants.name                     | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| representatives.identifier          | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| representatives.name                | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| applicationDate                     | \"==\", \"!=\", \"<\", \"<=\", \">\", \">=\"           |\n| registrationDate                    | \"==\", \"!=\", \"<\", \"<=\", \">\", \">=\"           |\n| designationDate                     | \"==\", \"!=\", \"<\", \"<=\", \">\", \">=\"           |\n| expiryDate                          | \"==\", \"!=\", \"<\", \"<=\", \">\", \">=\"           |\n| status                              | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| statusDate                          | \"==\", \"!=\", \"<\", \"<=\", \">\", \">=\"           |\n| description                         | \"==\", \"!=\"                                 |\n| niceClasses                         | \"==\", \"!=\", \"=in=\", \"=out=\", \"=all=\"       |\n| tradeDistinctivenessIndicator       | \"==\", \"!=\"                                 |\n| oppositionPeriodEndDate             | \"==\", \"!=\", \"<\", \"<=\", \">\", \">=\"           |\n| renewalStatus                       | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| fastTrackIndicator                  | \"==\", \"!=\"                                 |\n| oppositions.oppositionNumber        | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| cancellations.cancellationNumber    | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| appeals.appealNumber                | \"==\", \"!=\", \"=in=\", \"=out=\"                |\n| updateDate                          | \"<\", \"<=\", \">\", \">=\"                       |\n\n### Examples of valid queries\n- Trademarks from representative with identifier 123 for which renewal period is open\n```representatives.identifier==123 and renewalStatus==RENEWAL_PERIOD_OPEN```\n\n- Registered trademarks where their Goods and Services classification contains nice classes 25, 28 and 40 and verbal elements contains 'Dog'\n```niceClasses=all=(25,28,40) and wordMarkSpecification.verbalElement==*Dog* and status==REGISTERED```\n\n- Trademarks whose application date is greater than or equal to '2023-05-04' and they are FIGURATIVE and include nice classes 25,26 or they are WORD and they DON'T include nice class 40\n```applicationDate>=2023-05-04 and ((markFeature==FIGURATIVE and niceClasses=all=(25,26)) or (markFeature==WORD and niceClasses=out=(40)))```\n",
        "example": "applicationDate>=2023-05-04 and ((markFeature==FIGURATIVE and niceClasses=all=(25,26)) or (markFeature==WORD and niceClasses=out=(40)))"
      }
    },
    "responses": {
      "DefaultError": {
        "x-summary": "Unexpected error",
        "description": "Error occurred - see status code and problem object for more information.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "/problems/general-error",
              "title": "General error",
              "status": 500
            }
          }
        },
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/X-RateLimit-Remaining"
          },
          "X-RateLimit-Reset": {
            "$ref": "#/components/headers/X-RateLimit-Reset"
          },
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        }
      },
      "SearchTrademarkBadRequestError": {
        "description": "Bad request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "/problems/input-validation-error",
              "title": "Input validation error",
              "detail": "Query error: Unknown operator: =o=",
              "status": 400,
              "instance": "/trademarks"
            }
          }
        },
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/X-RateLimit-Remaining"
          },
          "X-RateLimit-Reset": {
            "$ref": "#/components/headers/X-RateLimit-Reset"
          },
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        }
      },
...
      "UnauthorizedError": {
        "x-summary": "Unauthorized",
        "description": "Client request has not been completed because it lacks valid authentication credentials for the requested resource",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "/problems/unauthorized",
              "title": "The request lacks valid authentication credentials for this resource",
              "detail": "Full authentication is required to access this resource",
              "status": 401,
              "instance": "/trademarks"
            }
          }
        },
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/X-RateLimit-Remaining"
          },
          "X-RateLimit-Reset": {
            "$ref": "#/components/headers/X-RateLimit-Reset"
          },
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        }
      },
      "RateLimitError": {
        "description": "Rate Limit Exceeded",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            },
            "example": {
              "type": "/problems/too-many-requests",
              "title": "Too Many Requests",
              "detail": "Rate Limit exceeded",
              "status": 429
            }
          }
        },
        "headers": {
          "X-RateLimit-Limit": {
            "$ref": "#/components/headers/X-RateLimit-Limit"
          },
          "X-RateLimit-Remaining": {
            "$ref": "#/components/headers/X-RateLimit-Remaining"
          },
          "X-RateLimit-Reset": {
            "$ref": "#/components/headers/X-RateLimit-Reset"
          },
          "Retry-After": {
            "$ref": "#/components/headers/Retry-After"
          }
        }
      }
    },
    "schemas": {
      "ApplicationNumber": {
        "type": "string",
        "pattern": "^(\\d{9}|W\\d{8})$",
        "maxLength": 9,
        "minLength": 9,
        "example": "018692868",
        "description": "Public identifier of the trade mark"
      },
      "PersonalReference": {
        "type": "string",
        "maxLength": 254,
        "pattern": "^[a-zA-Z0-9._\\\\-\\\\/\\\\\\\\]{1,254}$",
        "example": "PersonalReference001",
        "description": "Value that the user may indicate for his own reference. It should consist of up to 254 alphanumeric characters,\nusing the Roman alphabet: a-z, A-Z, 0-9 and/or the special characters: .-_/\\\\. Spaces not allowed\n"
      },
      "PageNumber": {
        "type": "integer",
        "format": "int32",
        "default": 0,
        "minimum": 0,
        "example": 0,
        "description": "The results are returned in pages. The page parameter (zero-based) indicates which page to return. For example, page = 2 (third page) and size = 10 will return matching applications No 21 to No 30"
      },
      "PageSize": {
        "type": "integer",
        "format": "int32",
        "default": 10,
        "minimum": 10,
        "maximum": 100,
        "example": 10,
        "description": "Number of results to fetch per page"
      },
      "TotalElements": {
        "type": "integer",
        "format": "int32",
        "minimum": 0,
        "example": 1,
        "description": "Total number of results"
      },
      "TotalPages": {
        "type": "integer",
        "format": "int32",
        "minimum": 0,
        "example": 1,
        "description": "Total number of pages"
      },
      "Sort": {
        "type": "string",
        "pattern": "^([\\w+]+(\\.(\\w)+)*:(asc|desc))$",
        "description": "Sorting",
        "maxLength": 50,
        "example": "applicationNumber:desc"
      },
...
      "TrademarkSearchResult": {
        "type": "object",
        "properties": {
          "trademarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrademarkSearchResultItem"
            },
            "maxItems": 100,
            "description": "List of trade marks that meet the search criteria"
          },
          "size": {
            "$ref": "#/components/schemas/PageSize"
          },
          "totalElements": {
            "$ref": "#/components/schemas/TotalElements"
          },
          "totalPages": {
            "$ref": "#/components/schemas/TotalPages"
          },
          "page": {
            "$ref": "#/components/schemas/PageNumber"
          }
        },
        "additionalProperties": false,
        "required": [
          "trademarks",
          "size",
          "totalElements",
          "totalPages",
          "page"
        ]
      },
      "TrademarkSearchResultItem": {
        "type": "object",
        "properties": {
          "applicationNumber": {
            "$ref": "#/components/schemas/ApplicationNumber"
          },
          "applicantReference": {
            "$ref": "#/components/schemas/PersonalReference"
          },
          "markKind": {
            "$ref": "#/components/schemas/MarkKind"
          },
          "markFeature": {
            "$ref": "#/components/schemas/MarkFeature"
          },
          "markBasis": {
            "$ref": "#/components/schemas/MarkBasis"
          },
          "niceClasses": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 45
            },
            "description": "Nice classification of the trade mark. The Nice Classification divides the goods and services\ninto 45 categories (or classes).\n",
            "externalDocs": {
              "description": "More information on Nice classification",
              "url": "https://euipo.europa.eu/ohimportal/en/nice-classification"
            }
          },
          "wordMarkSpecification": {
            "$ref": "#/components/schemas/WorkMarkSpecification"
          },
          "applicants": {
            "$ref": "#/components/schemas/ApplicantsList"
          },
          "representatives": {
            "$ref": "#/components/schemas/RepresentativesList"
          },
          "applicationDate": {
            "$ref": "#/components/schemas/ApplicationDate"
          },
          "registrationDate": {
            "$ref": "#/components/schemas/RegistrationDate"
          },
          "designationDate": {
            "$ref": "#/components/schemas/DesignationDate"
          },
          "expiryDate": {
            "$ref": "#/components/schemas/ExpiryDate"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        },
        "additionalProperties": false,
        "required": [
          "applicationNumber",
          "markFeature",
          "niceClasses",
          "status"
        ],
        "example": {
          "applicationNumber": "000984401",
          "markKind": "INDIVIDUAL",
          "markFeature": "WORD",
          "markBasis": "EU_TRADEMARK",
          "wordMarkSpecification": {
            "verbalElement": "White wine"
          },
          "niceClasses": [
            33
          ],
          "applicants": [
            {
              "office": "EM",
              "identifier": "1516"
            }
          ],
          "applicationDate": "1997-02-10",
          "registrationDate": "1998-10-28",
          "expiryDate": "2027-02-10",
          "status": "REGISTERED"
        }
      },
...
      "RepresentativesList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Person"
        },
        "description": "List of persons that represent the applicants of the trade mark"
      },
      "ApplicationDate": {
        "type": "string",
        "format": "date",
        "description": "Application date of the trade mark"
      },
...
      "MarkBasis": {
        "type": "string",
        "enum": [
          "EU_TRADEMARK",
          "INTERNATIONAL_TRADEMARK"
        ],
        "description": "Trademark mark basis\n\n- EU_TRADEMARK. European Union Trade Mark registration\n- INTERNATIONAL_TRADEMARK. International registration \n"
      },
...
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "description": "A URI reference that uniquely identifies the problem type only in the context of the provided API. Opposed to the specification in RFC-7807, it is neither recommended to be dereferenceable and point to a human-readable documentation nor globally unique for the problem type.\n",
            "default": "about:blank",
            "example": "/some/uri-reference"
          },
          "title": {
            "type": "string",
            "description": "A short summary of the problem type. Written in English and readable for engineers, usually not suited for non technical stakeholders and not localized.\n",
            "example": "some title for the error situation"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem.\n",
            "minimum": 100,
            "maximum": 600
          },
          "detail": {
            "type": "string",
            "description": "A human readable explanation specific to this occurrence of the problem that is helpful to locate the problem and give advice on how to proceed. Written in English and readable for engineers, usually not suited for non technical stakeholders and not localized.\n",
            "example": "some description for the error situation"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference",
            "description": "A URI reference that identifies the specific occurrence of the problem, e.g. by adding a fragment identifier or sub-path to the problem type. May be used to locate the root of this problem in the source code.\n",
            "example": "/some/uri-reference#specific-occurrence-context"
          }
        },
        "additionalProperties": false,
        "example": {
          "type": "/problems/not-found-error",
          "title": "Resource not found",
          "detail": "Application not found",
          "status": 404
        }
      }
    }
  },
...
}